@@ -1,4 +1,5 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
+import 'package:uevents/domain/user.dart';
 
 class Data {
   String uid;
@@ -11,13 +12,14 @@ class Data {
   Timestamp date;
   Timestamp startTime;
   Timestamp endTime;
-  
-  Data(this.uid,this.title, this.description, this.organizer, this.address,
-      this.shortDescription, this.imageSrc, this.date, this.startTime, this.endTime);
+  List<dynamic> participants;
+
+  Data(this.uid, this.title, this.description, this.organizer, this.address,
+      this.shortDescription, this.imageSrc, this.date, this.startTime, this.endTime, this.participants);
 
   Data copy() {
     return Data(this.uid, this.title, this.description, this.organizer,
-        this.address, this.shortDescription, this.imageSrc, this.date, this.startTime, this.endTime);
+        this.address, this.shortDescription, this.imageSrc, this.date, this.startTime, this.endTime, this.participants);
   }
 
   Data.fromJson(String uid, Map<String, dynamic> data){
@@ -31,6 +33,7 @@ class Data {
     date = data['date'];
     startTime = data['startTime'];
     endTime = data['endTime'];
+    participants = data['participants'];
   }
 
    Map<String, dynamic> toMap(){
@@ -45,6 +48,7 @@ class Data {
       "date": date,
       "startTime": startTime,
       "endTime": endTime,
+      "participants": participants
     };
   }
 }
